fix returns

+++ src/main/java/com/badals/flow/payload/RefundData.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public class RefundData {
    public boolean onUs;
-   public Double amount;
+   public BigDecimal amount;
    public int weight;
 }
